MAKE `StorageS3TestCase`
